requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
folium>=0.14.0
streamlit-folium>=0.13.0
plotly>=5.15.0
reportlab>=4.0.0
